Implement policies on bookshelves and others

+++ app/Livewire/Bookshelves/Edit.php
@@ -4,6 +4,8 @@ namespace App\Livewire\Bookshelves;
 
 use App\Livewire\Forms\BookshelfForm;
 use App\Models\Bookshelf;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Livewire\Attributes\Title;
 use Livewire\Component;
 
@@ -33,13 +35,16 @@ class Edit extends Component
     public function mount(string $bookshelf_number): void
     {
         $bookshelf = Bookshelf::where('bookshelf_number', $bookshelf_number)->firstOrFail();
-
+        $this->form->setBookshelf($bookshelf);
         if (!$bookshelf) {
             abort(404);
         }
+        if(Gate::denies('view', $this->form->bookshelf)) {
+            abort(403,'This action is unauthorized.');
+        }
         
         $this->user = $bookshelf->user->username;
-        $this->form->setBookshelf($bookshelf);
+        
     }
 
     /**
@@ -49,6 +54,9 @@ class Edit extends Component
      */
     public function save(): void
     {
+        if(Gate::denies('update', $this->form->bookshelf)) {
+            abort(403,'This action is unauthorized.');
+        }
         $this->form->update();
         $this->redirectRoute('bookshelves');
     }
@@ -60,10 +68,16 @@ class Edit extends Component
      */
     public function delete(): void
     {
-        $bookshelf = $this->form->bookshelf;
+        if(Gate::denies('update', $this->form->bookshelf)) {
+            abort(403,'This action is unauthorized.');
+        }
+        $bookshelf = Bookshelf::find($this->form->bookshelf->id);
         $bookshelf->delete();
-        session()->flash('success', 'Bookshelf deleted successfully');
-        $this->redirectRoute('bookshelves');
+        $bookshelf->update(['user_id' => Auth::user()->id]);
+            
+        session()->flash('success', 'Bookshelf successfully deleted');
+    
+        $this->dispatch('closeModal');
     }
 
     /**
